refactor: made card grid not have any width/height declaration, as it can be used in other bits
