+++ display/components/break/card_grid.module.css
@@ -1,21 +1,5 @@
-.main {
-    display: flex;
-    width: 1700px;
-    height: 850px;
-
-    align-items: center;
-    justify-content: center;
-    
-
-}
-
 .grid {
     display: grid;
     grid-template-columns: repeat(var(--columns), 1fr);
     gap:10px;
-
-    align-items: center;
-    justify-content: start;
-
-    transform: scale(var(--scale))
 }
